chore: seperate hooks into different files, pass in args for easier test
